splitting out the meta data inference like vendor and brand name
